Fixed relatedwords in database and matched this in DAL (decimal)

+++ DAL/Models/RelatedWordList.cs
@@ -12,6 +12,6 @@ namespace DAL.Models
         [Key]
         public string term { get; set; }
 
-        public double rank { get; set; }
+        public decimal rank { get; set; }
     }
 }
